Switch to profiles/census + code cleanup
